+++ CHANGELOG.md
@@ -31,6 +31,9 @@ Line wrap the file at 100 chars.                                              Th
 #### Android
 - Added toggle for Split tunneling view to be able to show system apps
 
+#### Windows
+- Resolve symbolic links and junctions for excluded apps.
+
 ### Changed
 - Only use the account history file to store the last used account.
 - Update the out of time-view and new account-view to make it more user friendly.